+++ cli/rpm/const.go
@@ -48,7 +48,9 @@ const (
 	tagPayloadCompressor = 1125
 	tagPayloadFlags      = 1126
 	tagPrein             = 1023
+	tagPostin            = 1024
 	tagPreinProg         = 1085
+	tagPostinProg        = 1086
 	tagDirNames          = 1118
 	tagBaseNames         = 1117
 	tagDirIndexes        = 1116